fix refactor management endpoint utils
